requirements.txt
playwright>=1.49
python-dotenv==1.0.1
